Follow-up: viewModel for YearInReview
