+++ app/src/main/java/org/wikipedia/yearinreview/YearInReviewViewModel.kt
@@ -23,37 +23,39 @@ import java.time.LocalDateTime
 import java.time.ZoneOffset
 
 class YearInReviewViewModel() : ViewModel() {
-    private val wikiAppContext = WikipediaApp.instance
-    private var _uiScreenListState = MutableStateFlow(Resource<List<YearInReviewScreenData>>())
-    val uiScreenListState: StateFlow<Resource<List<YearInReviewScreenData>>> = _uiScreenListState.asStateFlow()
-
-    val currentYear = LocalDate.now().year
-    val startTime: Instant = LocalDateTime.of(currentYear, 1, 1, 0, 0, 0).toInstant(ZoneOffset.UTC)
-    val endTime: Instant = LocalDateTime.of(currentYear, 12, 31, 23, 59, 59).toInstant(ZoneOffset.UTC)
-    val startTimeInMillis = startTime.toEpochMilli()
-    val endTimeInMillis = endTime.toEpochMilli()
-
-    val handler = CoroutineExceptionHandler { _, throwable ->
+    private val currentYear = LocalDate.now().year
+    private val startTime: Instant = LocalDateTime.of(currentYear, 1, 1, 0, 0, 0).toInstant(ZoneOffset.UTC)
+    private val endTime: Instant = LocalDateTime.of(currentYear, 12, 31, 23, 59, 59).toInstant(ZoneOffset.UTC)
+    private val startTimeInMillis = startTime.toEpochMilli()
+    private val endTimeInMillis = endTime.toEpochMilli()
+    private val handler = CoroutineExceptionHandler { _, throwable ->
         L.e(throwable)
         _uiScreenListState.value = Resource.Success(
             data = listOf(nonEnglishCollectiveReadCountData, nonEnglishCollectiveEditCountData)
         )
+    }
+    private var _uiScreenListState = MutableStateFlow(Resource<List<YearInReviewScreenData>>())
+    val uiScreenListState: StateFlow<Resource<List<YearInReviewScreenData>>> = _uiScreenListState.asStateFlow()
+
+    init {
+        fetchPersonalizedData()
     }
 
     fun fetchPersonalizedData() {
         val personalizedStatistics = YearInReviewStatistics()
         viewModelScope.launch(handler) {
             _uiScreenListState.value = Resource.Loading()
+
             val readCountJob = async {
                 personalizedStatistics.readCount = AppDatabase.instance.historyEntryDao().getHistoryCount(startTimeInMillis, endTimeInMillis)
-                if (personalizedStatistics.readCount >= 3) {
+                if (personalizedStatistics.readCount >= MINIMUM_READ_COUNT) {
                     personalizedStatistics.readCountApiTitles = AppDatabase.instance.historyEntryDao().getDisplayTitles()
                         .map { StringUtil.fromHtml(it).toString() }
-                    readCountData.headLineText = wikiAppContext.getString(
+                    readCountData.headLineText = WikipediaApp.instance.getString(
                         R.string.year_in_review_read_count_headline,
                         personalizedStatistics.readCount.toString()
                     )
-                    readCountData.bodyText = wikiAppContext.getString(
+                    readCountData.bodyText = WikipediaApp.instance.getString(
                         R.string.year_in_review_read_count_bodytext,
                         personalizedStatistics.readCount.toString(),
                         personalizedStatistics.readCountApiTitles[0],
@@ -65,57 +67,87 @@ class YearInReviewViewModel() : ViewModel() {
                     nonEnglishCollectiveReadCountData
                 }
             }
-            val editCountJob = async {
-                val homeSiteCall = async {
-                    ServiceFactory.get(WikipediaApp.instance.wikiSite)
-                        .getUserContribsByTimeFrame(
-                            username = AccountUtil.userName,
-                            maxCount = 500,
-                            startDate = endTime,
-                            endDate = startTime
-                        )
-                }
-                val commonsCall = async {
-                    ServiceFactory.get(Constants.commonsWikiSite)
-                        .getUserContribsByTimeFrame(
-                            username = AccountUtil.userName,
-                            maxCount = 500,
-                            startDate = endTime,
-                            endDate = startTime
-                        )
-                }
-                val wikidataCall = async {
-                    ServiceFactory.get(Constants.wikidataWikiSite)
-                        .getUserContribsByTimeFrame(
-                            username = AccountUtil.userName,
-                            maxCount = 500,
-                            startDate = endTime,
-                            endDate = startTime,
-                            ns = 0,
-                        )
-                }
 
-                val homeSiteResponse = homeSiteCall.await()
-                val commonsResponse = commonsCall.await()
-                val wikidataResponse = wikidataCall.await()
-
-                personalizedStatistics.editCount += homeSiteResponse.query?.userInfo!!.editCount
-                personalizedStatistics.editCount += wikidataResponse.query?.userInfo!!.editCount
-                personalizedStatistics.editCount += commonsResponse.query?.userInfo!!.editCount
-
-                editCountData.headLineText = wikiAppContext.getString(
-                    R.string.year_in_review_edit_count_headline,
-                    personalizedStatistics.editCount.toString()
-                )
-                editCountData.bodyText = wikiAppContext.getString(
-                    R.string.year_in_review_edit_count_bodytext,
-                    personalizedStatistics.editCount.toString()
-                )
-                editCountData
+            //
+            val homeSiteCall = async {
+                ServiceFactory.get(WikipediaApp.instance.wikiSite)
+                    .getUserContribsByTimeFrame(
+                        username = AccountUtil.userName,
+                        maxCount = 500,
+                        startDate = endTime,
+                        endDate = startTime
+                    )
             }
+            val commonsCall = async {
+                ServiceFactory.get(Constants.commonsWikiSite)
+                    .getUserContribsByTimeFrame(
+                        username = AccountUtil.userName,
+                        maxCount = 500,
+                        startDate = endTime,
+                        endDate = startTime
+                    )
+            }
+            val wikidataCall = async {
+                ServiceFactory.get(Constants.wikidataWikiSite)
+                    .getUserContribsByTimeFrame(
+                        username = AccountUtil.userName,
+                        maxCount = 500,
+                        startDate = endTime,
+                        endDate = startTime,
+                        ns = 0,
+                    )
+            }
+
+            val homeSiteResponse = homeSiteCall.await()
+            val commonsResponse = commonsCall.await()
+            val wikidataResponse = wikidataCall.await()
+
+            personalizedStatistics.editCount += homeSiteResponse.query?.userInfo!!.editCount
+            personalizedStatistics.editCount += wikidataResponse.query?.userInfo!!.editCount
+            personalizedStatistics.editCount += commonsResponse.query?.userInfo!!.editCount
+
+            editCountData.headLineText = WikipediaApp.instance.getString(
+                R.string.year_in_review_edit_count_headline,
+                personalizedStatistics.editCount.toString()
+            )
+            editCountData.bodyText = WikipediaApp.instance.getString(
+                R.string.year_in_review_edit_count_bodytext,
+                personalizedStatistics.editCount.toString()
+            )
+
             _uiScreenListState.value = Resource.Success(
-                data = listOf(readCountJob.await(), editCountJob.await())
+                data = listOf(readCountJob.await(), editCountData)
             )
         }
+    }
+
+    companion object {
+        private const val MINIMUM_READ_COUNT = 3
+
+        val getStartedData = YearInReviewScreenData(
+            imageResource = R.drawable.year_in_review_block_10_resize,
+            headLineText = R.string.year_in_review_get_started_headline,
+            bodyText = R.string.year_in_review_get_started_bodytext,
+        )
+
+        val readCountData = YearInReviewScreenData(
+            imageResource = R.drawable.wyir_block_5_resize
+        )
+
+        val editCountData = YearInReviewScreenData(
+            imageResource = R.drawable.wyir_bytes
+        )
+
+        val nonEnglishCollectiveReadCountData = YearInReviewScreenData(
+            imageResource = R.drawable.wyir_puzzle_3,
+            headLineText = R.string.year_in_review_non_english_collective_readcount_headline,
+            bodyText = R.string.year_in_review_non_english_collective_readcount_bodytext,
+        )
+
+        val nonEnglishCollectiveEditCountData = YearInReviewScreenData(
+            imageResource = R.drawable.wyir_puzzle_2_v5,
+            headLineText = R.string.year_in_review_non_english_collective_editcount_headline,
+            bodyText = R.string.year_in_review_non_english_collective_editcount_bodytext,
+        )
     }
 }
